issue-57: Remove Payout from Admin Panel and add filtering in rooms for admins
